Add tests and fix rubocop violations
